Fixed some channel mode bugs and a revenge bug.  Also added FounderMode.
git-svn-id: http://svn.neuromancy.net/magick/trunk/Magick-II@1410 1d68a2a4-f605-0410-a2a0-863aa388aa54

+++ src/ircsocket.cpp
@@ -676,8 +676,7 @@ void IrcSvcHandler::enqueue(mMessage * mm)
 	}
 
 	// Only spawn if we are less than our maximum ... and need it :)
-	if (message_queue.method_count() > static_cast < int >
-	    (tm.count_threads() * Magick::instance().config.High_Water_Mark()))
+	if (message_queue.method_count() > tm.count_threads() * Magick::instance().config.High_Water_Mark())
 	{
 	    CP(("Queue is full - Starting new thread and increasing watermarks ..."));
 	    if (static_cast < unsigned int > (tm.count_threads()) >= Magick::instance().config.Max_Threads())
